LayoutTests/imported/w3c/web-platform-tests/webcodecs/video-encoder.https.any-expected.txt


PASS Test VideoEncoder construction
PASS Test VideoEncoder.configure()
FAIL Test successful configure(), encode(), and flush() promise_test: Unhandled rejection with value: object "TypeError: undefined is not an object (evaluating 'decoderConfig.colorSpace.primaries')"
FAIL encodeQueueSize test assert_equals: expected 0 but got Infinity
PASS Test successful reset() and re-confiugre()
PASS Test successful encode() after re-configure().
PASS Verify closed VideoEncoder operations
PASS Verify unconfigured VideoEncoder operations
FAIL Verify encoding closed frames throws. assert_throws_dom: function "() => {
    encoder.encode(frame);
  }" threw object "TypeError: VideoFrame is detached" that is not a DOMException OperationError: property "code" is equal to undefined, expected 0
PASS Encode video with negative timestamp
